+++ src/main/java/com/amashchenko/maven/plugin/gitflow/GitFlowHotfixFinishMojo.java
@@ -26,7 +26,6 @@ import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.MojoFailureException;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
-import org.apache.maven.shared.release.versions.VersionParseException;
 import org.codehaus.plexus.components.interactivity.PrompterException;
 import org.codehaus.plexus.util.StringUtils;
 import org.codehaus.plexus.util.cli.CommandLineException;
@@ -62,46 +61,58 @@ public class GitFlowHotfixFinishMojo extends AbstractGitFlowMojo {
     @Parameter(property = "pushRemote", defaultValue = "true")
     private boolean pushRemote;
 
+    /**
+     * Maven goals to execute in the hotfix branch before merging into the
+     * production or support branch.
+     * 
+     * @since 1.8.0
+     */
+    @Parameter(property = "preHotfixGoals")
+    private String preHotfixGoals;
+
+    /**
+     * Maven goals to execute in the release or support branch after the hotfix.
+     * 
+     * @since 1.8.0
+     */
+    @Parameter(property = "postHotfixGoals")
+    private String postHotfixGoals;
+
+    /**
+     * Hotfix version to use in non interactive mode.
+     * 
+     */
+    @Parameter(property = "hotfixVersion")
+    private String hotfixVersion;
+
+    /**
+     * Whether to make a GPG-signed tag.
+     * 
+     */
+    @Parameter(property = "gpgSignTag", defaultValue = "false")
+    private boolean gpgSignTag = false;
+
     /** {@inheritDoc} */
     @Override
     public void execute() throws MojoExecutionException, MojoFailureException {
+        validateConfiguration(preHotfixGoals, postHotfixGoals);
+
         try {
             // check uncommitted changes
             checkUncommittedChanges();
 
-            // git for-each-ref --format='%(refname:short)' refs/heads/hotfix/*
-            final String hotfixBranches = gitFindBranches(
-                    gitFlowConfig.getHotfixBranchPrefix(), false);
-
-            if (StringUtils.isBlank(hotfixBranches)) {
-                throw new MojoFailureException("There are no hotfix branches.");
-            }
-
-            String[] branches = hotfixBranches.split("\\r?\\n");
-
-            List<String> numberedList = new ArrayList<String>();
-            StringBuilder str = new StringBuilder("Hotfix branches:")
-                    .append(LS);
-            for (int i = 0; i < branches.length; i++) {
-                str.append((i + 1) + ". " + branches[i] + LS);
-                numberedList.add(String.valueOf(i + 1));
-            }
-            str.append("Choose hotfix branch to finish");
-
-            String hotfixNumber = null;
-            try {
-                while (StringUtils.isBlank(hotfixNumber)) {
-                    hotfixNumber = prompter
-                            .prompt(str.toString(), numberedList);
-                }
-            } catch (PrompterException e) {
-                getLog().error(e);
-            }
-
             String hotfixBranchName = null;
-            if (hotfixNumber != null) {
-                int num = Integer.parseInt(hotfixNumber);
-                hotfixBranchName = branches[num - 1];
+            if (settings.isInteractiveMode()) {
+                hotfixBranchName = promptBranchName();
+            } else if (StringUtils.isNotBlank(hotfixVersion)) {
+                final String branch = gitFlowConfig.getHotfixBranchPrefix()
+                        + hotfixVersion;
+                if (!gitCheckBranchExists(branch)) {
+                    throw new MojoFailureException(
+                            "Hotfix branch with name '" + branch
+                                    + "' doesn't exist. Cannot finish hotfix.");
+                }
+                hotfixBranchName = branch;
             }
 
             if (StringUtils.isBlank(hotfixBranchName)) {
@@ -111,13 +122,13 @@ public class GitFlowHotfixFinishMojo extends AbstractGitFlowMojo {
 
             // support branch hotfix
             String supportBranchName = null;
-            boolean supportHotfix = hotfixBranchName.startsWith(gitFlowConfig
-                    .getHotfixBranchPrefix()
-                    + gitFlowConfig.getSupportBranchPrefix());
+            boolean supportHotfix = hotfixBranchName
+                    .startsWith(gitFlowConfig.getHotfixBranchPrefix()
+                            + gitFlowConfig.getSupportBranchPrefix());
             // get support branch name w/o version part
             if (supportHotfix) {
-                supportBranchName = hotfixBranchName.substring(gitFlowConfig
-                        .getHotfixBranchPrefix().length());
+                supportBranchName = hotfixBranchName.substring(
+                        gitFlowConfig.getHotfixBranchPrefix().length());
                 supportBranchName = supportBranchName.substring(0,
                         supportBranchName.lastIndexOf('/'));
             }
@@ -130,11 +141,11 @@ public class GitFlowHotfixFinishMojo extends AbstractGitFlowMojo {
                     gitFetchRemoteAndCompare(supportBranchName);
                 } else {
                     if (notSameProdDevName()) {
-                        gitFetchRemoteAndCompare(gitFlowConfig
-                                .getDevelopmentBranch());
+                        gitFetchRemoteAndCompare(
+                                gitFlowConfig.getDevelopmentBranch());
                     }
-                    gitFetchRemoteAndCompare(gitFlowConfig
-                            .getProductionBranch());
+                    gitFetchRemoteAndCompare(
+                            gitFlowConfig.getProductionBranch());
                 }
             }
 
@@ -144,6 +155,13 @@ public class GitFlowHotfixFinishMojo extends AbstractGitFlowMojo {
 
                 // mvn clean test
                 mvnCleanTest();
+            }
+
+            // maven goals before merge
+            if (StringUtils.isNotBlank(preHotfixGoals)) {
+                gitCheckout(hotfixBranchName);
+
+                mvnRun(preHotfixGoals);
             }
 
             if (supportBranchName != null) {
@@ -156,17 +174,22 @@ public class GitFlowHotfixFinishMojo extends AbstractGitFlowMojo {
             // git merge --no-ff hotfix/...
             gitMergeNoff(hotfixBranchName);
 
-            final String hotfixVersion = getCurrentProjectVersion();
+            final String currentVersion = getCurrentProjectVersion();
             if (!skipTag) {
-                String tagVersion = hotfixVersion;
+                String tagVersion = currentVersion;
                 if (tychoBuild && ArtifactUtils.isSnapshot(tagVersion)) {
-                    tagVersion = tagVersion.replace("-"
-                            + Artifact.SNAPSHOT_VERSION, "");
+                    tagVersion = tagVersion
+                            .replace("-" + Artifact.SNAPSHOT_VERSION, "");
                 }
 
                 // git tag -a ...
                 gitTag(gitFlowConfig.getVersionTagPrefix() + tagVersion,
-                        commitMessages.getTagHotfixMessage());
+                        commitMessages.getTagHotfixMessage(), gpgSignTag);
+            }
+
+            // maven goals after merge
+            if (StringUtils.isNotBlank(postHotfixGoals)) {
+                mvnRun(postHotfixGoals);
             }
 
             // check whether release branch exists
@@ -184,7 +207,7 @@ public class GitFlowHotfixFinishMojo extends AbstractGitFlowMojo {
                     gitMergeNoff(hotfixBranchName);
                 } else {
                     GitFlowVersionInfo developVersionInfo = new GitFlowVersionInfo(
-                            hotfixVersion);
+                            currentVersion);
                     if (notSameProdDevName()) {
                         // git checkout develop
                         gitCheckout(gitFlowConfig.getDevelopmentBranch());
@@ -193,7 +216,7 @@ public class GitFlowHotfixFinishMojo extends AbstractGitFlowMojo {
                                 getCurrentProjectVersion());
 
                         // set version to avoid merge conflict
-                        mvnSetVersions(hotfixVersion);
+                        mvnSetVersions(currentVersion);
                         gitCommit("update to hotfix version");
 
                         // git merge --no-ff hotfix/...
@@ -201,8 +224,9 @@ public class GitFlowHotfixFinishMojo extends AbstractGitFlowMojo {
 
                         // which version to increment
                         GitFlowVersionInfo hotfixVersionInfo = new GitFlowVersionInfo(
-                                hotfixVersion);
-                        if (developVersionInfo.compareTo(hotfixVersionInfo) < 0) {
+                                currentVersion);
+                        if (developVersionInfo
+                                .compareTo(hotfixVersionInfo) < 0) {
                             developVersionInfo = hotfixVersionInfo;
                         }
                     }
@@ -256,10 +280,50 @@ public class GitFlowHotfixFinishMojo extends AbstractGitFlowMojo {
                     gitPushDelete(hotfixBranchName);
                 }
             }
-        } catch (CommandLineException e) {
-            getLog().error(e);
-        } catch (VersionParseException e) {
-            getLog().error(e);
+        } catch (Exception e) {
+            throw new MojoFailureException("hotfix-finish", e);
         }
+    }
+
+    private String promptBranchName() throws MojoFailureException, CommandLineException {
+        // git for-each-ref --format='%(refname:short)' refs/heads/hotfix/*
+        String hotfixBranches = gitFindBranches(gitFlowConfig.getHotfixBranchPrefix(), false);
+
+        // find hotfix support branches
+        if (!gitFlowConfig.getHotfixBranchPrefix().endsWith("/")) {
+            String supportHotfixBranches = gitFindBranches(gitFlowConfig.getHotfixBranchPrefix() + "*/*", false);
+            hotfixBranches = hotfixBranches + supportHotfixBranches;
+        }
+
+        if (StringUtils.isBlank(hotfixBranches)) {
+            throw new MojoFailureException("There are no hotfix branches.");
+        }
+
+        String[] branches = hotfixBranches.split("\\r?\\n");
+
+        List<String> numberedList = new ArrayList<String>();
+        StringBuilder str = new StringBuilder("Hotfix branches:").append(LS);
+        for (int i = 0; i < branches.length; i++) {
+            str.append((i + 1) + ". " + branches[i] + LS);
+            numberedList.add(String.valueOf(i + 1));
+        }
+        str.append("Choose hotfix branch to finish");
+
+        String hotfixNumber = null;
+        try {
+            while (StringUtils.isBlank(hotfixNumber)) {
+                hotfixNumber = prompter.prompt(str.toString(), numberedList);
+            }
+        } catch (PrompterException e) {
+            throw new MojoFailureException("hotfix-finish", e);
+        }
+
+        String hotfixBranchName = null;
+        if (hotfixNumber != null) {
+            int num = Integer.parseInt(hotfixNumber);
+            hotfixBranchName = branches[num - 1];
+        }
+
+        return hotfixBranchName;
     }
 }