+++ src/main/java/com/amashchenko/maven/plugin/gitflow/GitFlowReleaseMojo.java
@@ -24,10 +24,8 @@ import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.MojoFailureException;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
-import org.apache.maven.shared.release.versions.VersionParseException;
 import org.codehaus.plexus.components.interactivity.PrompterException;
 import org.codehaus.plexus.util.StringUtils;
-import org.codehaus.plexus.util.cli.CommandLineException;
 
 /**
  * The git flow release mojo.
@@ -125,9 +123,34 @@ public class GitFlowReleaseMojo extends AbstractGitFlowMojo {
     @Parameter(property = "versionDigitToIncrement")
     private Integer versionDigitToIncrement;
 
+    /**
+     * Maven goals to execute before the release.
+     * 
+     * @since 1.8.0
+     */
+    @Parameter(property = "preReleaseGoals")
+    private String preReleaseGoals;
+
+    /**
+     * Maven goals to execute after the release.
+     * 
+     * @since 1.8.0
+     */
+    @Parameter(property = "postReleaseGoals")
+    private String postReleaseGoals;
+
+    /**
+     * Whether to make a GPG-signed tag.
+     * 
+     */
+    @Parameter(property = "gpgSignTag", defaultValue = "false")
+    private boolean gpgSignTag = false;
+
     /** {@inheritDoc} */
     @Override
     public void execute() throws MojoExecutionException, MojoFailureException {
+        validateConfiguration(preReleaseGoals, postReleaseGoals);
+
         try {
             // set git flow configuration
             initGitFlowConfig();
@@ -207,14 +230,20 @@ public class GitFlowReleaseMojo extends AbstractGitFlowMojo {
                         }
                     }
                 } catch (PrompterException e) {
-                    getLog().error(e);
+                    throw new MojoFailureException("release", e);
                 }
             } else {
                 version = releaseVersion;
             }
 
             if (StringUtils.isBlank(version)) {
+                getLog().info("Version is blank. Using default version.");
                 version = defaultVersion;
+            }
+
+            // maven goals before release
+            if (StringUtils.isNotBlank(preReleaseGoals)) {
+                mvnRun(preReleaseGoals);
             }
 
             // execute if version changed
@@ -245,7 +274,12 @@ public class GitFlowReleaseMojo extends AbstractGitFlowMojo {
 
                 // git tag -a ...
                 gitTag(gitFlowConfig.getVersionTagPrefix() + version,
-                        commitMessages.getTagReleaseMessage());
+                        commitMessages.getTagReleaseMessage(), gpgSignTag);
+            }
+
+            // maven goals after release
+            if (StringUtils.isNotBlank(postReleaseGoals)) {
+                mvnRun(postReleaseGoals);
             }
 
             if (notSameProdDevName()) {
@@ -293,10 +327,8 @@ public class GitFlowReleaseMojo extends AbstractGitFlowMojo {
                     gitPush(gitFlowConfig.getDevelopmentBranch(), !skipTag);
                 }
             }
-        } catch (CommandLineException e) {
-            getLog().error(e);
-        } catch (VersionParseException e) {
-            getLog().error(e);
+        } catch (Exception e) {
+            throw new MojoFailureException("release", e);
         }
     }
 }